docs: add example for resampling bugs: fix small bugs in import test: update tests for test_valid (new file) and test_process
